* moved tests out of unnecessarily deeply nested folder structure * increased coverage in mergeCartsOnLoginTest
