+++ extension/test/unit/bigcommerce/AuthRepositoryTest.js
@@ -1,7 +1,7 @@
 'use strict'
 const assert = require('assert')
 const jwt = require('jwt-simple')
-const AuthRepository = require('../../../../../lib/bigcommerce/AuthRepository')
+const AuthRepository = require('../../../lib/bigcommerce/AuthRepository')
 
 describe('BigCommerceAuthRepository - unit', () => {
   const clientId = '1111111'
