+++ extension/test/unit/bigcommerce/CartRepositoryTest.js
@@ -3,9 +3,9 @@ const BigCommerce = require('node-bigcommerce')
 const sinon = require('sinon')
 const assert = require('assert')
 const chai = require('chai')
-const BigCommerceCartRepository = require('../../../../../lib/bigcommerce/CartRepository')
-const BigCommerceCart = require('../../../../../lib/bigcommerce/Cart')
-const BigCommerceCartLineItemRequest = require('../../../../../lib/bigcommerce/cart/LineItemRequest')
+const BigCommerceCartRepository = require('../../../lib/bigcommerce/CartRepository')
+const BigCommerceCart = require('../../../lib/bigcommerce/Cart')
+const BigCommerceCartLineItemRequest = require('../../../lib/bigcommerce/cart/LineItemRequest')
 
 chai.use(require('chai-subset'))
 chai.use(require('chai-as-promised')).should()
