+++ extension/test/unit/bigcommerce/cart/LineItemsTest.js
@@ -1,10 +1,10 @@
 'use strict'
 const assert = require('assert')
 const chai = require('chai')
-const BigCommerceCartLineItems = require('../../../../../../lib/bigcommerce/cart/LineItems')
-const BigCommerceCartLineItemPhysical = require('../../../../../../lib/bigcommerce/cart/line_item/Physical')
-const BigCommerceCartLineItemDigital = require('../../../../../../lib/bigcommerce/cart/line_item/Digital')
-const BigCommerceCartLineItemGiftCertificate = require('../../../../../../lib/bigcommerce/cart/line_item/GiftCertificate')
+const BigCommerceCartLineItems = require('../../../../lib/bigcommerce/cart/LineItems')
+const BigCommerceCartLineItemPhysical = require('../../../../lib/bigcommerce/cart/line_item/Physical')
+const BigCommerceCartLineItemDigital = require('../../../../lib/bigcommerce/cart/line_item/Digital')
+const BigCommerceCartLineItemGiftCertificate = require('../../../../lib/bigcommerce/cart/line_item/GiftCertificate')
 
 chai.use(require('chai-subset'))
 chai.use(require('chai-as-promised')).should()
