+++ extension/test/unit/shopgate/CartExtensionPipelineTest.js
@@ -4,10 +4,10 @@ const sinon = require('sinon')
 const assert = require('assert')
 const chai = require('chai')
 const {describe, it, beforeEach, afterEach} = require('mocha')
-const ShopgateCartExtensionPipeline = require('../../../../../lib/shopgate/CartExtensionPipeline')
-const IdentifierConverter = require('../../../../../lib/shopgate/IdentifierConverter')
-const ShopgateCartFactory = require('../../../../../lib/shopgate/CartFactory')
-const BigCommerceCartRepository = require('../../../../../lib/bigcommerce/CartRepository')
+const ShopgateCartExtensionPipeline = require('../../../lib/shopgate/CartExtensionPipeline')
+const IdentifierConverter = require('../../../lib/shopgate/IdentifierConverter')
+const ShopgateCartFactory = require('../../../lib/shopgate/CartFactory')
+const BigCommerceCartRepository = require('../../../lib/bigcommerce/CartRepository')
 
 chai.use(require('chai-subset'))
 chai.use(require('chai-as-promised')).should()
