+++ extension/test/unit/shopgate/IdentifierConverterTest.js
@@ -1,7 +1,7 @@
 'use strict'
 const assert = require('assert')
 const {describe, it, beforeEach} = require('mocha')
-const IdentifierConverter = require('../../../../../lib/shopgate/IdentifierConverter')
+const IdentifierConverter = require('../../../lib/shopgate/IdentifierConverter')
 
 describe('CartExtensionPipeline - unit', () => {
   /** @type IdentifierConverter */
